+++ stdlib/generic/.formwork/fixtures/go-weak-types-any/fire-1/pkg/api.go
@@ -1,0 +1,5 @@
+//go:build ignore
+
+package pkg
+
+func Handle(x any) {} // want: go-weak-types-any