+++ stdlib/generic/.formwork/fixtures/go-weak-types-reflect/fire-1/pkg/api.go
@@ -1,0 +1,9 @@
+//go:build ignore
+
+package pkg
+
+import "reflect"
+
+func NameOf(v any) string {
+	return reflect.TypeOf(v).String() // want: go-weak-types-reflect
+}