+++ stdlib/generic/.formwork/fixtures/todo-fixme-hack-dart/fire-1/lib/app.dart
@@ -1,0 +1,2 @@
+// TODO: wire the screen // want: todo-fixme-hack-dart
+void main() {}